backend/requirements.txt
fastapi[standard]==0.116.1
sqlalchemy
alembic
psycopg2-binary
pylint
pyhumps
passlib
python-jose
redis
loguru
bcrypt
pytest